fix(adapters): Use correct volume_step instead of hardcoding 1.0 (#158)
* fix(adapters): Use correct volume_step instead of hardcoding 1.0

BREAKING: VolumeInfo now includes step field

Fixes regression where all adapters hardcoded volume_step to 1.0,
ignoring backend-specific values.

Changes by adapter:

- **Roon**: Use v.step from API (varies per zone)
  - Added step: Option<f32> to VolumeInfo struct
  - Map step from roon-api Volume in parse_zone()
  - Use v.step.unwrap_or(1.0) in roon_zone_to_bus_zone()

- **LMS**: Use constant 2.5 (from LMS server)
  - LMS hardcodes $increment = 2.5 in Slim/Player/Client.pm:755
  - Not queryable via CLI/JSON-RPC, so we use the constant

- **OpenHome**: Query Characteristics action for VolumeSteps
  - Added volume_max and volume_steps to OpenHomeDevice
  - Poll Characteristics action during device update
  - Calculate step = VolumeMax / VolumeSteps

- **HQPlayer**: Already correct (no changes needed)

Closes: https://github.com/cloud-atlas-ai/unified-hifi-control/issues/152
Related: https://github.com/muness/roon-knob/issues/102


* fix(api): Include volume_control in /zones endpoint

The /zones endpoint was returning volume_step: null because ZoneInfo
didn't include the volume_control field from Zone. Added the field
and mapping to expose step to clients.

Also added lint tests to prevent regression.


* fix(ui): Show decimal volumes when step is fractional

The web UI was always rounding volume to integers. Now it shows
one decimal place when volume_step is fractional (e.g., 0.5 for
Roon, 2.5 for LMS).

Also added volume_step to NowPlaying API type.


* fix(volume): Support fractional volume steps throughout the stack

Changed volume step handling from i32 to f32 to preserve fractional
precision (e.g., 0.5 dB for Roon, 2.5% for LMS).

Changes:
- Roon adapter: change_volume takes f32, rounds at transport layer
- LMS adapter: change_volume takes f32, rounds at control layer
- Routes: Remove as i32 casts, keep step as f32 throughout
- API: VolumeRequest/LmsVolumeRequest use f32 for value
- Tests: Updated to use f32 literals

The Roon/LMS transport APIs still use i32, so rounding happens at
the last moment to maximize precision through the stack.


* fix(volume): Look up zone's actual step when no value provided

The web UI sends vol_up/vol_down without a step value. Previously
this defaulted to 1.0, ignoring the zone's configured step.

Now the control handlers look up the zone's volume_control.step
from the aggregator when no value is provided.


* fix(aggregator): Update zone.now_playing on track changes

Bug: NowPlayingChanged updated a separate HashMap, but routes read from
zone.now_playing which was only set on initial ZoneDiscovered. This caused
track info (title, artist, album) to become stale.

Fix: Remove the separate now_playing HashMap. Update zone.now_playing
directly when NowPlayingChanged arrives. Single source of truth.


* fix: Web UI volume buttons use zone's step instead of hardcoded 2

- Change ControlRequest.value from i32 to f64 for fractional steps
- Extract volume_step from now_playing data
- Volume +/- buttons now use zone's step (e.g., 0.5 dB for Roon)


* fix: Upgrade roon-api to 0.3.1 fork with fractional volume support

- Switch to open-horizon-labs fork with f64 volume fix
- Update CoreEvent::Found -> CoreEvent::Registered for 0.3.1 API
- Remove i32 rounding in change_volume call

Fixes fractional dB steps (e.g., 0.5 dB) being rounded to 1 dB.

Upstream PR: https://github.com/TheAppgineer/rust-roon-api/pull/1


---------

+++ src/api/mod.rs
@@ -285,11 +285,11 @@ pub async fn roon_control_handler(
     }
 }
 
-/// Volume request body
+/// Volume request body (f32 for fractional step support)
 #[derive(Deserialize)]
 pub struct VolumeRequest {
     pub output_id: String,
-    pub value: i32,
+    pub value: f32,
     #[serde(default)]
     pub relative: bool,
 }
@@ -704,7 +704,7 @@ pub async fn lms_control_handler(
 #[derive(Deserialize)]
 pub struct LmsVolumeRequest {
     pub player_id: String,
-    pub value: i32,
+    pub value: f32,
     #[serde(default)]
     pub relative: bool,
 }
